requirements.txt
Django==1.6.1
Markdown==2.3.1
dj-database-url==0.2.2
django-model-utils==2.0
factory-boy==2.3.1
gunicorn==18.0
hurl==2.1
psycopg2==2.5.2
wheel==0.22.0
